+++ actix-tls/src/connect/resolver.rs
@@ -1,3 +1,4 @@
+use core::future::{ready, Ready};
 use std::{
     future::Future,
     io,
@@ -10,7 +11,6 @@ use std::{
 
 use actix_rt::task::{spawn_blocking, JoinHandle};
 use actix_service::{Service, ServiceFactory};
-use actix_utils::future::{ok, Ready};
 use futures_core::{future::LocalBoxFuture, ready};
 use tracing::trace;
 
@@ -45,7 +45,7 @@ impl<R: Host> ServiceFactory<ConnectInfo<R>> for Resolver {
     type Future = Ready<Result<Self::Service, Self::InitError>>;
 
     fn new_service(&self, _: ()) -> Self::Future {
-        ok(self.resolver.clone())
+        ready(Ok(self.resolver.clone()))
     }
 }