+++ actix-tls/src/connect/rustls_0_22.rs
@@ -2,6 +2,7 @@
 //!
 //! See [`TlsConnector`] for main connector service factory docs.
 
+use core::future::{ready, Ready};
 use std::{
     future::Future,
     io,
@@ -12,7 +13,6 @@ use std::{
 
 use actix_rt::net::ActixStream;
 use actix_service::{Service, ServiceFactory};
-use actix_utils::future::{ok, Ready};
 use futures_core::ready;
 use rustls_pki_types_1::ServerName;
 use tokio_rustls::{
@@ -91,9 +91,9 @@ where
     type Future = Ready<Result<Self::Service, Self::InitError>>;
 
     fn new_service(&self, _: ()) -> Self::Future {
-        ok(TlsConnectorService {
+        ready(Ok(TlsConnectorService {
             connector: self.connector.clone(),
-        })
+        }))
     }
 }