+++ actix-tls/tests/accept-openssl.rs
@@ -7,6 +7,7 @@
     feature = "openssl"
 ))]
 
+use core::future::ready;
 use std::{io::Write as _, sync::Arc};
 
 use actix_rt::net::TcpStream;
@@ -16,7 +17,6 @@ use actix_tls::{
     accept::openssl::{Acceptor, TlsStream},
     connect::rustls_0_23::reexports::ClientConfig,
 };
-use actix_utils::future::ok;
 use rustls_pki_types_1::ServerName;
 use tokio_rustls_026::rustls::RootCertStore;
 
@@ -127,7 +127,7 @@ async fn accepts_connections() {
 
             tls_acceptor
                 .map_err(|err| println!("OpenSSL error: {err:?}"))
-                .and_then(move |_stream: TlsStream<TcpStream>| ok(()))
+                .and_then(move |_stream: TlsStream<TcpStream>| ready(Ok(())))
         }
     });