+++ actix-server/CHANGES.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-- Minimum supported Rust version (MSRV) is now 1.76.
+- Minimum supported Rust version (MSRV) is now 1.88.
+- internal: Use `core::future::{ready, Ready}` instead of actix-utils'
 
 ## 2.6.0